+++ src/main/java/app/repo/InCardRepo.java
@@ -3,5 +3,5 @@ package app.repo;
 import app.entity.InCard;
 import org.springframework.data.repository.CrudRepository;
 
-public interface InCardRepo extends CrudRepository<InCard,Long> {
+public interface InCardRepo extends CrudRepository<InCard,Integer> {
 }
