+++ src/main/java/app/repo/SauceRepo.java
@@ -3,5 +3,5 @@ package app.repo;
 import app.entity.Sauce;
 import org.springframework.data.repository.CrudRepository;
 
-public interface SauceRepo extends CrudRepository<Sauce,Long> {
+public interface SauceRepo extends CrudRepository<Sauce,Integer> {
 }
