+++ src/main/java/app/service/IngredientService.java
@@ -1,6 +1,5 @@
 package app.service;
 
-import app.entity.Ingredient;
 import app.repo.IngredientRepo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -20,7 +19,7 @@ public class IngredientService {
     }
 
     private void deleteIngredientByID(int id) {
-        ingredientRepo.deleteById((long) id);
+        ingredientRepo.deleteById(id);
     }
 
     private void deleteIngredient(Ingredient ingredient) {
@@ -28,7 +27,7 @@ public class IngredientService {
     }
 
     private Ingredient getIngredient(int id) {
-        Optional<Ingredient> optIngredient = ingredientRepo.findById((long) id);
+        Optional<Ingredient> optIngredient = ingredientRepo.findById(id);
         return optIngredient.orElseGet(Ingredient::new);
     }
